pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Brownfield Cartographer"
requires-python = ">=3.11"
dependencies = [
    "tree-sitter",
    "pyyaml",
    "pydantic",
    "networkx",
    "gitpython",
    "tree-sitter-languages>=1.10.2",
    "sqlglot>=29.0.1",
]
